add in speed time plus edits
attempted to reverse classification starting with feeding/travelling but classification becomes more difficult as you go down so gave up and returned to the resting/travelling split.
Added in speed time model (on the basis of looking back 30 days). designed to be run along a moving window.
changed movement distance in roosting to be 15m (from 25m)