Add ability to have moving 30 day window
create `day30window` variable in data frame and allocate data to a window.  use this variable as an interaction term.  If the full model fits, also try an interaction model and use BIC to choose which is better.
Also adjust `plot_model_fit` to allow multiple lines if interaction is selected.